Comments API Finalization
